move topology out
