+++ client/session.go
@@ -22,7 +22,6 @@ package client
 
 import (
 	"errors"
-	"fmt"
 	"math"
 	"sort"
 	"sync"
@@ -240,194 +239,7 @@ func (s *session) Open() error {
 	return nil
 }
 
-func (s *session) initHostQueues(topoMap topology.Map) ([]hostQueue, int, int, error) {
-	// NB(r): we leave existing writes in the host queues to finish
-	// as they are already enroute to their destination, this is ok
-	// as part of adding a host is to add another replica for the
-	// shard set and only once bootstrapped decomission the old node
-	start := s.nowFn()
-
-	hosts := topoMap.Hosts()
-	queues := make([]hostQueue, len(hosts))
-	for i := range queues {
-		queues[i] = s.newHostQueue(hosts[i], topoMap)
-	}
-
-	shards := topoMap.ShardSet().AllIDs()
-	minConnectionCount := s.opts.MinConnectionCount()
-	replicas := topoMap.Replicas()
-	majority := topoMap.MajorityReplicas()
-
-	firstConnectConsistencyLevel := s.opts.ClusterConnectConsistencyLevel()
-	if firstConnectConsistencyLevel == ConnectConsistencyLevelNone {
-		// Return immediately if no connect consistency required
-		return queues, replicas, majority, nil
-	}
-
-	connectConsistencyLevel := firstConnectConsistencyLevel
-	if connectConsistencyLevel == ConnectConsistencyLevelAny {
-		// If level any specified, first attempt all then proceed lowering requirement
-		connectConsistencyLevel = ConnectConsistencyLevelAll
-	}
-
-	// Abort if we do not connect
-	connected := false
-	defer func() {
-		if !connected {
-			for _, q := range queues {
-				q.Close()
-			}
-		}
-	}()
-
-	for {
-		if now := s.nowFn(); now.Sub(start) >= s.opts.ClusterConnectTimeout() {
-			switch firstConnectConsistencyLevel {
-			case ConnectConsistencyLevelAny:
-				// If connecting with connect any strategy then keep
-				// trying but lower consistency requirement
-				start = now
-				connectConsistencyLevel--
-				if connectConsistencyLevel == ConnectConsistencyLevelNone {
-					// Already tried to resolve all consistency requirements, just
-					// return successfully at this point
-					return queues, replicas, majority, nil
-				}
-			default:
-				// Timed out connecting to a specific consistency requirement
-				return nil, 0, 0, ErrClusterConnectTimeout
-			}
-		}
-		// Be optimistic
-		clusterAvailable := true
-		for _, shard := range shards {
-			shardReplicasAvailable := 0
-			routeErr := topoMap.RouteShardForEach(shard, func(idx int, _ topology.Host) {
-				if queues[idx].ConnectionCount() >= minConnectionCount {
-					shardReplicasAvailable++
-				}
-			})
-			if routeErr != nil {
-				return nil, 0, 0, routeErr
-			}
-			var clusterAvailableForShard bool
-			switch connectConsistencyLevel {
-			case ConnectConsistencyLevelAll:
-				clusterAvailableForShard = shardReplicasAvailable == replicas
-			case ConnectConsistencyLevelMajority:
-				clusterAvailableForShard = shardReplicasAvailable >= majority
-			case ConnectConsistencyLevelOne:
-				clusterAvailableForShard = shardReplicasAvailable > 0
-			default:
-				return nil, 0, 0, errSessionInvalidConnectClusterConnectConsistencyLevel
-			}
-			if !clusterAvailableForShard {
-				clusterAvailable = false
-				break
-			}
-		}
-		if clusterAvailable {
-			// All done
-			break
-		}
-		time.Sleep(clusterConnectWaitInterval)
-	}
-
-	connected = true
-	return queues, replicas, majority, nil
-}
-
-func (s *session) setTopologyWithLock(topologyMap topology.Map, hasLock bool) error {
-	queues, replicas, majority, err := s.initHostQueues(topologyMap)
-	if err != nil {
-		return err
-	}
-
-	if !hasLock {
-		s.Lock()
-		defer s.Unlock()
-	}
-
-	prev := s.queues
-	s.queues = queues
-	s.topoMap = topoMap
-
-	prevReplicas := atomic.LoadInt32(&s.replicas)
-	atomic.StoreInt32(&s.replicas, int32(replicas))
-	atomic.StoreInt32(&s.majority, int32(majority))
-
-	if s.fetchBatchOpArrayArrayPool == nil ||
-		s.fetchBatchOpArrayArrayPool.Entries() != len(queues) {
-		poolOpts := pool.NewObjectPoolOptions().
-			SetSize(s.opts.FetchBatchOpPoolSize()).
-			SetInstrumentOptions(s.opts.InstrumentOptions().SetMetricsScope(
-				s.scope.SubScope("fetch-batch-op-array-array-pool"),
-			))
-		s.fetchBatchOpArrayArrayPool = newFetchBatchOpArrayArrayPool(
-			poolOpts,
-			len(queues),
-			s.opts.FetchBatchOpPoolSize()/len(queues))
-		s.fetchBatchOpArrayArrayPool.Init()
-	}
-	if s.iteratorArrayPool == nil ||
-		prevReplicas != s.replicas {
-		s.iteratorArrayPool = encoding.NewIteratorArrayPool([]pool.Bucket{
-			pool.Bucket{
-				Capacity: replicas,
-				Count:    s.opts.SeriesIteratorPoolSize(),
-			},
-		})
-		s.iteratorArrayPool.Init()
-	}
-	if s.readerSliceOfSlicesIteratorPool == nil ||
-		prevReplicas != s.replicas {
-		size := replicas * s.opts.SeriesIteratorPoolSize()
-		poolOpts := pool.NewObjectPoolOptions().
-			SetSize(size).
-			SetInstrumentOptions(s.opts.InstrumentOptions().SetMetricsScope(
-				s.scope.SubScope("reader-slice-of-slices-iterator-pool"),
-			))
-		s.readerSliceOfSlicesIteratorPool = newReaderSliceOfSlicesIteratorPool(poolOpts)
-		s.readerSliceOfSlicesIteratorPool.Init()
-	}
-	if s.multiReaderIteratorPool == nil ||
-		prevReplicas != s.replicas {
-		size := replicas * s.opts.SeriesIteratorPoolSize()
-		poolOpts := pool.NewObjectPoolOptions().
-			SetSize(size).
-			SetInstrumentOptions(s.opts.InstrumentOptions().SetMetricsScope(
-				s.scope.SubScope("multi-reader-iterator-pool"),
-			))
-		s.multiReaderIteratorPool = encoding.NewMultiReaderIteratorPool(poolOpts)
-		s.multiReaderIteratorPool.Init(s.opts.ReaderIteratorAllocate())
-	}
-
-	for i := len(s.metrics.writeNodesRespondingErrors); i < replicas; i++ {
-		tags := map[string]string{"nodes": fmt.Sprintf("%d", i+1)}
-		counter := s.scope.Tagged(tags).Counter("write.nodes-responding-error")
-		s.metrics.writeNodesRespondingErrors =
-			append(s.metrics.writeNodesRespondingErrors, counter)
-	}
-
-	for i := len(s.metrics.fetchNodesRespondingErrors); i < replicas; i++ {
-		tags := map[string]string{"nodes": fmt.Sprintf("%d", i+1)}
-		counter := s.scope.Tagged(tags).Counter("fetch.nodes-responding-error")
-		s.metrics.fetchNodesRespondingErrors =
-			append(s.metrics.fetchNodesRespondingErrors, counter)
-	}
-
-	// Asynchronously close the previous set of host queues
-	go func() {
-		for _, p := range prev {
-			p.Close()
-		}
-	}()
-
-	s.log.Infof("successfully updated topology to %d hosts", topologyMap.HostsLen())
-	return nil
-}
-
-func (s *session) newHostQueue(host topology.Host, topoMap topology.Map) hostQueue {
+func (s *session) newHostQueue(host topology.Host, topologyMap topology.Map) hostQueue {
 	// NB(r): Due to hosts being replicas we have:
 	// = replica * numWrites
 	// = total writes to all hosts
